Move /react/src/utils/** to /react/src/** - Move loose components to /react/src/components/ - Delete unused upload/FilesTable.tsx
