+++ react/src/admin/CreateUserModal.tsx
@@ -8,7 +8,7 @@ import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import TextField from "@material-ui/core/TextField";
-import { NewUser } from "../utils/typings";
+import { NewUser } from "../typings";
 
 export interface CreateUserModalProps {
     id: string;
