+++ react/src/admin/ManagedUploaders.tsx
@@ -5,7 +5,7 @@ import MaterialTable, { MTableToolbar } from "material-table";
 import HidableTextButton from "./HidableTextButton";
 import Grid from "@material-ui/core/Grid";
 import Paper from "@material-ui/core/Paper";
-import Card from "../Card";
+import Card from "../components/Card";
 
 const useStyles = makeStyles(theme => ({
     chip: {
