+++ react/src/analysis/SetAssigneeDialog.tsx
@@ -9,7 +9,7 @@ import {
     DialogTitle,
     TextField,
 } from "@material-ui/core";
-import { Analysis } from "../utils/typings";
+import { Analysis } from "../typings";
 
 interface SetAssigneeDialogProps {
     affectedRows: Analysis[];
