+++ react/src/components/AnalysisInfoDialog.tsx
