+++ react/src/components/BooleanDisplay.tsx
