+++ react/src/components/BooleanEditComponent.tsx
