+++ react/src/components/BooleanFilter.tsx
