+++ react/src/components/DetailSection.tsx
