+++ react/src/components/GridFieldsDisplay.tsx
