+++ react/src/components/Note.tsx
