+++ react/src/components/SecretDisplay.tsx
